backend/requirements.txt
# AcquiSmart Backend Dependencies

# ===== Web Framework =====
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ===== Validation & Config =====
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ===== Database =====
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
opensearch-py>=2.4.0

# ===== Auth (Cognito) =====
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# ===== Payments =====
stripe>=7.0.0

# ===== Storage =====
aioboto3>=12.0.0
cryptography>=41.0.0

# ===== OCR & Documents =====
# Note: paddleocr requires additional system deps
# paddleocr>=2.7.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pdfplumber>=0.10.0
reportlab>=4.0.0
PyPDF2>=3.0.0

# ===== ML/NLP =====
# Note: torch is large, install separately if needed
# torch>=2.1.0
transformers>=4.36.0
spacy>=3.7.0
sentence-transformers>=2.2.0

# ===== Entity Resolution =====
rapidfuzz>=3.5.0
# faiss-cpu>=1.7.4  # Install separately

# ===== LLM Integration =====
openai>=1.6.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# ===== Analytics & Recommendations =====
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.1.0

# ===== Testing =====
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ===== Development =====
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
